% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{bar_plot}
\alias{bar_plot}
\title{Plota grafico de barras estatico dos precos das marmitas}
\usage{
bar_plot()
}
\value{
Um grafico de barras estatico
}
\description{
\code{bar_plot()} exibe um grafico de barras estatico em que e possivel observar
os valores das marmitas em ordem decrescente.
}
\examples{
bar_plot()
}
