% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotly.R
\name{bar_plotly}
\alias{bar_plotly}
\title{Plota um grafico de barras iterativo dos precos das marmitas}
\usage{
bar_plotly()
}
\value{
Um grafico de barras iterativo
}
\description{
\code{bar_plotly()} exibe um grafico dinamico em que e possivel observar
os valores das marmitas em ordem decrescente.
}
\examples{
bar_plotly()
}
