% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marmitas_veganas}
\alias{marmitas_veganas}
\title{Dados sobre marmitas veganas do Beleaf 100\% vegetal e sem lactose}
\format{
Uma tibble, contendo:
\describe{
\item{item}{ID de um item do cardapio}
\item{nome}{Nome de um item do cardapio}
\item{descricao}{Descricao de um item do cardapio}
\item{preco}{Preco de um item do cardapio em Reais}
\item{tipo_de_refeicao}{Tipo de refeicao}
}
}
\source{
Beleaf - \url{https://www.beleaf.com.br/cardapio-marmitas-veganas}
}
\usage{
marmitas_veganas
}
\description{
Essa base de dados contem informacoes sobre as marmitas veganas do Beleaf
}
\examples{
head(marmitas_veganas)
}
\keyword{datasets}
