% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunplotly.R
\name{sun_plotly}
\alias{sun_plotly}
\title{Plota um grafico de sol iterativo com os precos das marmitas}
\usage{
sun_plotly()
}
\value{
Um grafico de sol iterativo
}
\description{
\code{sun_plotly()} e uma funcao que retorna um grafico iterativo em que e
possivel clicar nas categorias para exibi-las com mais detalhes.
}
\examples{
sun_plotly()
}
